testing frame time

+++ FishMain/Fish.h
@@ -8,11 +8,11 @@
      //String fishName;
     float GetFalseBitPercentage(int currentCatchChance);
      String fishName;
-     float GetBiteStrength() {return biteStrength;}
+     unsigned long GetBiteStrength() {return biteStrength;}
      int points;
 
 private:
-     int biteStrength;
+     unsigned long biteStrength;
      float falseBitePercentage; // false bite percentage is based on the current catch percentage (for example, falseBitePercentage(30 / 100) * (100 - currentCatchChance))
      float GetFishValue() {return points;}
     //  float GetBiteStrength() {return biteStrength;}
